# api/openapi.yaml
openapi: 3.1.0
info:
  title: dra-workflow API
  version: 0.1.0
  summary: Workflow-driven, AI-assisted IT service disaster recovery planning and execution.
  description: |
    REST API of the dra-workflow backend (Rust). It implements the data model in
    `docs/domain/dr-plan-model.md` and the workflows in `docs/workflows/`.

    **Hierarchy:** Tenant → IT service → microservice → DR items.

    **Methodology:** NIST SP 800-34 Rev. 1 and BSI-Standard 200-4 (see `docs/domain/standards-mapping.md`).

    ## Conventions
    - **Tenant from the token:** the tenant is **not** part of the path. The backend reads it from the
      `tenant_id` claim of the JWT access token and scopes every query to it. A token without a valid
      tenant claim, or for a tenant the user is not a member of, gets `403`. Ids belonging to another
      tenant return `404`. Only `GET/POST /tenants` (membership list, tenant creation) are tenant-independent.
    - **Collections** are nested under their parent (for example `/services/{serviceId}/microservices`).
      **Single items** are addressed flat by id (for example `/microservices/{microserviceId}`).
    - IDs are UUIDs. Durations are **integer minutes**. Timestamps are RFC 3339 UTC.
    - Updates use `PATCH` with `application/merge-patch+json` (RFC 7396).
    - Optimistic concurrency: every resource has a `version`. Reads return an `ETag`, and `PATCH`/`DELETE`
      accept `If-Match`. A stale version returns `412`.
...
      with `issues`.
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /members:
    get:
      tags: [Directory]
      operationId: listMembers
      summary: List tenant members (application users)
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Members
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MemberPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Directory]
      operationId: addMember
      summary: Add a member with a tenant role
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MemberCreate' }
      responses:
        '201':
          description: Member added
          headers:
            Location: { $ref: '#/components/headers/Location' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Member' }
        '409': { $ref: '#/components/responses/Conflict' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }
  /members/{memberId}:
    parameters:
      - $ref: '#/components/parameters/MemberId'
    patch:
      tags: [Directory]
      operationId: updateMember
      summary: Change a member's tenant role
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { $ref: '#/components/schemas/MemberUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Member' }
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }
    delete:
      tags: [Directory]
      operationId: removeMember
      summary: Remove a member
      responses:
        '204': { description: Removed }
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }
  /persons:
    get:
      tags: [Directory]
      operationId: listPersons
      summary: List persons (contact directory)
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: q
          in: query
          description: Search by name, email or team
          schema: { type: string }
      responses:
        '200':
          description: Persons
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PersonPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Directory]
      operationId: createPerson
      summary: Create a person
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PersonCreate' }
      responses:
        '201':
          description: Created
          headers:
            Location: { $ref: '#/components/headers/Location' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Person' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }
  /persons/{personId}:
    parameters:
      - $ref: '#/components/parameters/PersonId'
    get:
      tags: [Directory]
      operationId: getPerson
      summary: Get a person
      responses:
        '200':
          description: Person
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Person' }
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }
    patch:
      tags: [Directory]
      operationId: updatePerson
      summary: Update a person
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { $ref: '#/components/schemas/PersonFields' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Person' }
        '404': { $ref: '#/components/responses/NotFound' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }
    delete:
      tags: [Directory]
      operationId: deletePerson
      summary: Delete a person
      description: Fails with `409` while the person is still assigned to a role or owns open action items.
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }
  /roles:
    get:
      tags: [Directory]
      operationId: listRoles
      summary: List the DR role catalog (Incident Commander, DBA, and others)
      responses:
        '200':
          description: Roles
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Role' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Directory]
      operationId: createRole
      summary: Create a custom DR role
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RoleCreate' }
      responses:
        '201':
          description: Created
          headers:
            Location: { $ref: '#/components/headers/Location' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Role' }
        '409': { $ref: '#/components/responses/Conflict' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }
  /roles/{roleId}:
    parameters:
      - $ref: '#/components/parameters/RoleId'
    patch:
      tags: [Directory]
      operationId: updateRole
      summary: Update a role
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { $ref: '#/components/schemas/RoleFields' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Role' }
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }
    delete:
      tags: [Directory]
      operationId: deleteRole
      summary: Delete a custom role
      description: Fails with `409` while the role is referenced by assignments or runbook steps.
      responses:
        '204': { description: Deleted }
        '409': { $ref: '#/components/responses/Conflict' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }

  # ───────────────────────────── Catalog ─────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        OIDC access token (JWT). Required claims: `sub` (user) and `tenant_id` (UUID of the active tenant).
        The tenant role may come from the token or from the tenant membership. The identity provider is
        still open (docs/overview/open-questions.md).
  headers:
    Location:
      description: URL of the created resource
      schema: { type: string }
...
  parameters:
...
    MemberId:
      name: memberId
      in: path
      required: true
      schema: { type: string, format: uuid }
    PersonId:
      name: personId
      in: path
      required: true
      schema: { type: string, format: uuid }
    RoleId:
      name: roleId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor from a previous page's `nextCursor`
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    IfMatch:
      name: If-Match
      in: header
      description: ETag of the version being modified (optimistic concurrency)
      schema: { type: string }
...
  responses:
    Unauthorized:
      description: Missing or invalid access token
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Error:
      description: Unexpected error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource not found (or it belongs to a different tenant than the token's)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Conflict:
      description: Conflicts with the current state (for example the resource is still referenced, or the transition is invalid)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    PreconditionFailed:
      description: The `If-Match` version is stale
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ValidationFailed:
      description: Input validation, integrity rule or workflow gate failed
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ValidationProblem' }
...
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 problem details
      properties:
        type: { type: string, format: uri-reference, default: about:blank }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
    ValidationProblem:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          properties:
            issues:
              type: array
              items: { $ref: '#/components/schemas/Issue' }
    Issue:
      type: object
      description: A validation error, integrity-rule violation or gate failure
      required: [ruleId, severity, message]
      properties:
        ruleId:
          type: string
          description: Stable rule id, for example `RTO_EXCEEDS_SERVICE_RTO`, `DEPENDENCY_RTO_CONFLICT`, `STEP_WITHOUT_OWNER`, `STRATEGY_GAP`, `ROLE_WITHOUT_DEPUTY`
          examples: [DEPENDENCY_RTO_CONFLICT]
        severity: { type: string, enum: [blocking, warning, info] }
        message: { type: string }
        field: { type: string, description: 'JSON pointer to the offending input field, if any' }
        entityType: { type: string }
        entityId: { type: string, format: uuid }
        workflowStep: { $ref: '#/components/schemas/WorkflowStepKey' }
        standardRef:
          type: string
          description: Related standard reference
          examples: ['NIST SP 800-34 §3.2', 'BSI 200-4 Soll-Ist-Vergleich']
    ResourceMeta:
      type: object
      required: [id, tenantId, version, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid, readOnly: true }
        tenantId: { type: string, format: uuid, readOnly: true }
        version: { type: integer, readOnly: true, description: Incremented on every change (optimistic concurrency) }
        createdAt: { type: string, format: date-time, readOnly: true }
        createdBy: { type: string, readOnly: true }
        updatedAt: { type: string, format: date-time, readOnly: true }
        updatedBy: { type: string, readOnly: true }
...
    PageInfo:
      type: object
      properties:
        nextCursor: { type: [string, 'null'] }
...
    TenantRole:
      type: string
      enum: [admin, author, reviewer, responder, auditor]
...
    WorkflowStepKey:
      type: string
      enum:
        - define_service
        - map_dependencies
        - business_impact
        - brainstorm_scenarios
        - consolidate_scenarios
        - select_scenarios
        - recovery_objectives
        - recovery_strategies
        - runbooks
        - roles
        - communication
        - review_approve
        - test
        - measure
        - improve
...
    MemberCreate:
      type: object
      required: [userRef, tenantRole]
      properties:
        userRef: { type: string, description: Subject identifier from the identity provider }
        email: { type: string, format: email }
        displayName: { type: string }
        tenantRole: { $ref: '#/components/schemas/TenantRole' }
        personId: { type: [string, 'null'], format: uuid, description: Optional link to a contact directory entry }
    MemberUpdate:
      type: object
      properties:
        tenantRole: { $ref: '#/components/schemas/TenantRole' }
        personId: { type: [string, 'null'], format: uuid }
    Member:
      allOf:
        - $ref: '#/components/schemas/ResourceMeta'
        - $ref: '#/components/schemas/MemberCreate'
    MemberPage:
      allOf:
        - $ref: '#/components/schemas/PageInfo'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Member' }
    PersonFields:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        email: { type: string, format: email }
        phone: { type: string }
        alternateContact: { type: string, description: Out-of-band contact for use when normal channels are down }
        team: { type: string }
    PersonCreate:
      allOf:
        - $ref: '#/components/schemas/PersonFields'
        - required: [name]
    Person:
      allOf:
        - $ref: '#/components/schemas/ResourceMeta'
        - $ref: '#/components/schemas/PersonFields'
        - required: [name]
    PersonPage:
      allOf:
        - $ref: '#/components/schemas/PageInfo'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Person' }
    RoleFields:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
    RoleCreate:
      allOf:
        - $ref: '#/components/schemas/RoleFields'
        - required: [name]
    Role:
      allOf:
        - $ref: '#/components/schemas/ResourceMeta'
        - $ref: '#/components/schemas/RoleFields'
        - type: object
          required: [name]
          properties:
            isDefault: { type: boolean, readOnly: true }

    # ─────────── Catalog ───────────
...